solve compilation warning caused by comparing int with size_t in clock_connection_builders.cpp

+++ vpr/src/route/clock_connection_builders.cpp
@@ -255,7 +255,7 @@ void ClockToPinsConnection::create_switches(const ClockRRGraphBuilder& clock_gra
     for (int x = 0; x < (int)grid.width(); x++) {
         for (int y = 0; y < (int)grid.height(); y++) {
             //Avoid boundary
-            if ((y == 0 && x == 0) || (x == grid.width() - 1 && y == grid.height() - 1)) {
+            if ((y == 0 && x == 0) || (x == (int)grid.width() - 1 && y == (int)grid.height() - 1)) {
                 continue;
             }
 
@@ -285,7 +285,7 @@ void ClockToPinsConnection::create_switches(const ClockRRGraphBuilder& clock_gra
 
             for (e_side side : SIDES) {
                 //Don't connect pins which are not adjacent to channels around the perimeter
-                if ((x == 0 && side != RIGHT) || (x == grid.width() - 1 && side != LEFT) || (y == 0 && side != TOP) || (y == grid.height() - 1 && side != BOTTOM)) {
+                if ((x == 0 && side != RIGHT) || (x == (int)grid.width() - 1 && side != LEFT) || (y == 0 && side != TOP) || (y == (int)grid.height() - 1 && side != BOTTOM)) {
                     continue;
                 }
 
@@ -301,7 +301,7 @@ void ClockToPinsConnection::create_switches(const ClockRRGraphBuilder& clock_gra
                     if (x == 0) {
                         clock_x_offset = 1;  // chanx clock always starts at 1 offset
                         clock_y_offset = -1; // pick the chanx below the block
-                    } else if (x == grid.width() - 1) {
+                    } else if (x == (int)grid.width() - 1) {
                         clock_x_offset = -1; // chanx clock always ends at 1 offset
                         clock_y_offset = -1; // pick the chanx below the block
                     } else if (y == 0) {
